feat: update types for js modules
