+++ @convey/babel-plugin/index.d.ts
@@ -1,0 +1,3 @@
+import M from './main'
+
+export = M
