+++ @convey/next-plugin/index.d.ts
@@ -1,0 +1,3 @@
+import Module from './main'
+
+export = Module
