+++ @convey/vite-plugin/index.d.ts
