+++ @examples/nextjs/next.config.js
@@ -10,7 +10,6 @@ const withConvey = require('@convey/next-plugin')({
 	},
 })
 
-/** @param {import('next').NextConfig} nextConfig */
 module.exports = withConvey({
 	compress: false,
