+++ @examples/nextjs/package.json
@@ -14,7 +14,7 @@
 		"@examples/nextjs": "../nextjs",
 		"@prisma/client": "^6.2.1",
 		"devcert": "1.2.2",
-		"next": "^15.1.6",
+		"next": "*",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"superstruct": "^2.0.2"
